[moo-updateTip] show progress bar syn and reload new tip + delete distance at Detail + hide status bar at new photo

+++ NearHoneyTip/NHTMainTableCell.m
@@ -46,7 +46,7 @@
     self.userNickname.text = self.tip.userNickname;
     self.tipDate.text = self.tip.tipDate;
     
-    NSString *distanceWithKm = [NSString stringWithFormat:@"%u", self.tip.distance];
+    NSString *distanceWithKm = [NSString stringWithFormat:@"%lu", (unsigned long)self.tip.distance];
     distanceWithKm = [distanceWithKm stringByAppendingString:@" m"];
     
     self.distance.text = distanceWithKm;
@@ -55,6 +55,7 @@
     
     UITapGestureRecognizer *tapLikeButton = [[UITapGestureRecognizer alloc]initWithTarget:self action:@selector(didTapLike:)];
     self.likeButton.gestureRecognizers = [[NSArray alloc] initWithObjects: tapLikeButton, nil];
+    
   
     [self setReplyButtonProperty];
 }
@@ -72,7 +73,7 @@
     if (self.tip.likeInteger <= 0){
         likeCount = @"";
     } else {
-        likeCount = [NSString stringWithFormat:@"%u", self.tip.likeInteger];
+        likeCount = [NSString stringWithFormat:@"%lu", (unsigned long)self.tip.likeInteger];
     }
     likeString = [likeString stringByAppendingString:likeCount];
     [self.likeButton setTitle:likeString forState:UIControlStateNormal];
@@ -126,14 +127,13 @@
         [self.postManager postLikeChangeMethod:@"PUT" Tip:self.tip.tipId];
     }
     
-   // NHTMainViewController *refresh = [[NHTMainViewController alloc]init];
-   // [refresh getLatestTips];
+
 }
 
 -(void)willPlusLike{
     self.tip.likeInteger = self.tip.likeInteger + 1 ;
     NSString *likeString = @"좋아요 ";
-    likeString = [likeString stringByAppendingString: [NSString stringWithFormat:@"%u", self.tip.likeInteger]];
+    likeString = [likeString stringByAppendingString: [NSString stringWithFormat:@"%lu", (unsigned long)self.tip.likeInteger]];
     [self.likeButton setTitle:likeString forState:UIControlStateNormal];
     
     
@@ -142,7 +142,7 @@
     self.tip.likeInteger = self.tip.likeInteger - 1;
     NSString *likeString = @"좋아요 ";
     if(self.tip.likeInteger > 0){
-        likeString = [likeString stringByAppendingString: [NSString stringWithFormat:@"%u", self.tip.likeInteger]];
+        likeString = [likeString stringByAppendingString: [NSString stringWithFormat:@"%lu", (unsigned long)self.tip.likeInteger]];
     }
     [self.likeButton setTitle:likeString forState:UIControlStateNormal];
